fix: order migrations

+++ migrations/sqls/20221125200452-orders-table-up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE orders
 (
     id      SERIAL PRIMARY KEY,
-    user_id bigint REFERENCES users (id),
+    user_id INTEGER REFERENCES users (id),
     status  VARCHAR(32)
 );
