+++ migrations/sqls/20221125200513-orders-products-table-up.sql
@@ -1,7 +1,7 @@
 CREATE TABLE orders_products
 (
     id SERIAL PRIMARY KEY,
-    product_quantity integer,
-    order_id bigint REFERENCES orders(id),
-    product_id bigint REFERENCES products(id)
+    product_quantity INTEGER,
+    order_id INTEGER REFERENCES orders(id),
+    product_id INTEGER REFERENCES products(id)
 );
